+++ src/components/common/index.ts
@@ -3,6 +3,5 @@ export { default as ImageButton } from './ImageButton';
 export { default as ExpandTooltipButton } from './ExpandTooltipButton';
 export { default as SlideScroll } from './SlideScroll';
 export { default as FadeIn } from './FadeIn';
-export { default as SearchForm } from './SearchForm';
 export { default as QuestionCarousel } from './QuestionCarousel';
 export { default as AvatarCard } from './AvatarCard';
