+++ src/components/search/index.ts
@@ -1,1 +1,2 @@
 export { default as QuizletItem } from './QuizletItem';
+export { default as SearchForm } from './SearchForm';
